Refactor Commits models
